LPS-75081 Add support for CONCAT Function
Once this is committed, the SQLTransformer will be able to accept SQL statements using the CONCAT operator on an indefinite number of expressions.

For databases that only allow the CONCAT operator on two expressions, it will convert the call to a number of nested calls to CONCAT (e.g. "CONCAT('a', 'b', 'c', 'd')" would become "CONCAT('a', CONCAT('b', CONCAT('c', 'd')))").

For DBs that don't support the CONCAT function, it will replace the call to the CONCAT function with the arguments separated by the concatenation operator (e.g. "CONCAT('a', 'b', 'c', 'd')" would become "'a' + 'b' + 'c' + 'd'").