LPS-x Create a new method _processURLAsVirtualHostLayoutFriendlyURL.
The new method will take in a url and process it as though it were a URL
pointing to a layout on the virtual host (e.g. "/home"). It will return
true if the URL could be processed as pointing to a layout on the
virtual host, and false otherwise.