LPD-17038 Register upgrade process to create AnalyticsMessage table from 1.1.0 to 1.2.0, as it is in 7.3. This ensures that users can upgrade directly from 7.3 U30 to 7.4, instead of having to go through 7.3 U32 first. This effectively "forward-ports" the fix from LPS-193731 to 7.4.
This is necessary because it is not an expected behavior that users
would be required to upgrade to 7.3 U32 before they can upgrade to 7.4.
They should be able to upgrade to 7.4 directly from any 7.3 U level.

This should be safe because the upgrade process will check if the table
exists first before it tries to create it. So if they are upgrading from
an older 7.3 level, like 7.3 dxp-3 for example, the upgrade process will
not cause any failures because it will check first and notice that the
AnalyticsMessage table already exists, so it will not try to create it.